feat(ui): Weather & Banner are user controlled
Changes:
- Weather and Banner visibility can now be controlled by the user through the UI.
- New options appears in popup menu under the options menu.
- 2 options stored in the UI Properties
- Number of hotspots rendered is shown in statusbar
- Weather display now in a dedicated new class WeatherPanel.

ref #2077
